requirements.txt
python-dotenv
pytest
pytest-playwright
playwright
allure-pytest